Move flake8-bugbear violations to rule modules
