+++ src/registry.rs
@@ -99,34 +99,34 @@ ruff_macros::define_rule_mapping!(
     A002 => violations::BuiltinArgumentShadowing,
     A003 => violations::BuiltinAttributeShadowing,
     // flake8-bugbear
-    B002 => violations::UnaryPrefixIncrement,
-    B003 => violations::AssignmentToOsEnviron,
-    B004 => violations::UnreliableCallableCheck,
-    B005 => violations::StripWithMultiCharacters,
-    B006 => violations::MutableArgumentDefault,
-    B007 => violations::UnusedLoopControlVariable,
-    B008 => violations::FunctionCallArgumentDefault,
-    B009 => violations::GetAttrWithConstant,
-    B010 => violations::SetAttrWithConstant,
-    B011 => violations::DoNotAssertFalse,
-    B012 => violations::JumpStatementInFinally,
-    B013 => violations::RedundantTupleInExceptionHandler,
-    B014 => violations::DuplicateHandlerException,
-    B015 => violations::UselessComparison,
-    B016 => violations::CannotRaiseLiteral,
-    B017 => violations::NoAssertRaisesException,
-    B018 => violations::UselessExpression,
-    B019 => violations::CachedInstanceMethod,
-    B020 => violations::LoopVariableOverridesIterator,
-    B021 => violations::FStringDocstring,
-    B022 => violations::UselessContextlibSuppress,
-    B023 => violations::FunctionUsesLoopVariable,
-    B024 => violations::AbstractBaseClassWithoutAbstractMethod,
-    B025 => violations::DuplicateTryBlockException,
-    B026 => violations::StarArgUnpackingAfterKeywordArg,
-    B027 => violations::EmptyMethodWithoutAbstractDecorator,
-    B904 => violations::RaiseWithoutFromInsideExcept,
-    B905 => violations::ZipWithoutExplicitStrict,
+    B002 => rules::flake8_bugbear::rules::UnaryPrefixIncrement,
+    B003 => rules::flake8_bugbear::rules::AssignmentToOsEnviron,
+    B004 => rules::flake8_bugbear::rules::UnreliableCallableCheck,
+    B005 => rules::flake8_bugbear::rules::StripWithMultiCharacters,
+    B006 => rules::flake8_bugbear::rules::MutableArgumentDefault,
+    B007 => rules::flake8_bugbear::rules::UnusedLoopControlVariable,
+    B008 => rules::flake8_bugbear::rules::FunctionCallArgumentDefault,
+    B009 => rules::flake8_bugbear::rules::GetAttrWithConstant,
+    B010 => rules::flake8_bugbear::rules::SetAttrWithConstant,
+    B011 => rules::flake8_bugbear::rules::DoNotAssertFalse,
+    B012 => rules::flake8_bugbear::rules::JumpStatementInFinally,
+    B013 => rules::flake8_bugbear::rules::RedundantTupleInExceptionHandler,
+    B014 => rules::flake8_bugbear::rules::DuplicateHandlerException,
+    B015 => rules::flake8_bugbear::rules::UselessComparison,
+    B016 => rules::flake8_bugbear::rules::CannotRaiseLiteral,
+    B017 => rules::flake8_bugbear::rules::NoAssertRaisesException,
+    B018 => rules::flake8_bugbear::rules::UselessExpression,
+    B019 => rules::flake8_bugbear::rules::CachedInstanceMethod,
+    B020 => rules::flake8_bugbear::rules::LoopVariableOverridesIterator,
+    B021 => rules::flake8_bugbear::rules::FStringDocstring,
+    B022 => rules::flake8_bugbear::rules::UselessContextlibSuppress,
+    B023 => rules::flake8_bugbear::rules::FunctionUsesLoopVariable,
+    B024 => rules::flake8_bugbear::rules::AbstractBaseClassWithoutAbstractMethod,
+    B025 => rules::flake8_bugbear::rules::DuplicateTryBlockException,
+    B026 => rules::flake8_bugbear::rules::StarArgUnpackingAfterKeywordArg,
+    B027 => rules::flake8_bugbear::rules::EmptyMethodWithoutAbstractDecorator,
+    B904 => rules::flake8_bugbear::rules::RaiseWithoutFromInsideExcept,
+    B905 => rules::flake8_bugbear::rules::ZipWithoutExplicitStrict,
     // flake8-blind-except
     BLE001 => violations::BlindExcept,
     // flake8-comprehensions
