+++ src/rules/flake8_bugbear/rules/abstract_base_class.rs
@@ -1,10 +1,38 @@
-use rustpython_ast::{Constant, Expr, ExprKind, Keyword, Stmt, StmtKind};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::{Diagnostic, Rule};
-use crate::violations;
+use crate::violation::Violation;
 use crate::visibility::{is_abstract, is_overload};
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Constant, Expr, ExprKind, Keyword, Stmt, StmtKind};
+
+define_violation!(
+    pub struct AbstractBaseClassWithoutAbstractMethod {
+        pub name: String,
+    }
+);
+impl Violation for AbstractBaseClassWithoutAbstractMethod {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let AbstractBaseClassWithoutAbstractMethod { name } = self;
+        format!("`{name}` is an abstract base class, but it has no abstract methods")
+    }
+}
+define_violation!(
+    pub struct EmptyMethodWithoutAbstractDecorator {
+        pub name: String,
+    }
+);
+impl Violation for EmptyMethodWithoutAbstractDecorator {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let EmptyMethodWithoutAbstractDecorator { name } = self;
+        format!(
+            "`{name}` is an empty method in an abstract base class, but has no abstract decorator"
+        )
+    }
+}
 
 fn is_abc_class(checker: &Checker, bases: &[Expr], keywords: &[Keyword]) -> bool {
     keywords.iter().any(|keyword| {
@@ -91,7 +119,7 @@ pub fn abstract_base_class(
 
         if !has_abstract_decorator && is_empty_body(body) && !is_overload(checker, decorator_list) {
             checker.diagnostics.push(Diagnostic::new(
-                violations::EmptyMethodWithoutAbstractDecorator {
+                EmptyMethodWithoutAbstractDecorator {
                     name: format!("{name}.{method_name}"),
                 },
                 Range::from_located(stmt),
@@ -105,7 +133,7 @@ pub fn abstract_base_class(
     {
         if !has_abstract_method {
             checker.diagnostics.push(Diagnostic::new(
-                violations::AbstractBaseClassWithoutAbstractMethod {
+                AbstractBaseClassWithoutAbstractMethod {
                     name: name.to_string(),
                 },
                 Range::from_located(stmt),
