+++ src/rules/flake8_bugbear/rules/assert_false.rs
@@ -1,11 +1,26 @@
-use rustpython_ast::{Constant, Expr, ExprContext, ExprKind, Location, Stmt, StmtKind};
-
 use crate::ast::helpers::unparse_stmt;
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::fix::Fix;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::AlwaysAutofixableViolation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Constant, Expr, ExprContext, ExprKind, Location, Stmt, StmtKind};
+
+define_violation!(
+    pub struct DoNotAssertFalse;
+);
+impl AlwaysAutofixableViolation for DoNotAssertFalse {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Do not `assert False` (`python -O` removes these calls), raise `AssertionError()`")
+    }
+
+    fn autofix_title(&self) -> String {
+        "Replace `assert False`".to_string()
+    }
+}
 
 fn assertion_error(msg: Option<&Expr>) -> Stmt {
     Stmt::new(
@@ -46,7 +61,7 @@ pub fn assert_false(checker: &mut Checker, stmt: &Stmt, test: &Expr, msg: Option
         return;
     };
 
-    let mut diagnostic = Diagnostic::new(violations::DoNotAssertFalse, Range::from_located(test));
+    let mut diagnostic = Diagnostic::new(DoNotAssertFalse, Range::from_located(test));
     if checker.patch(diagnostic.kind.rule()) {
         diagnostic.amend(Fix::replacement(
             unparse_stmt(&assertion_error(msg), checker.stylist),
