+++ src/rules/flake8_bugbear/rules/assert_raises_exception.rs
@@ -7,13 +7,23 @@
 //! typo. Either assert for a more specific exception (builtin or
 //! custom), use `assertRaisesRegex`, or use the context manager form of
 //! `assertRaises`.
-
-use rustpython_ast::{ExprKind, Stmt, Withitem};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{ExprKind, Stmt, Withitem};
+
+define_violation!(
+    pub struct NoAssertRaisesException;
+);
+impl Violation for NoAssertRaisesException {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("`assertRaises(Exception)` should be considered evil")
+    }
+}
 
 /// B017
 pub fn assert_raises_exception(checker: &mut Checker, stmt: &Stmt, items: &[Withitem]) {
@@ -41,7 +51,7 @@ pub fn assert_raises_exception(checker: &mut Checker, stmt: &Stmt, items: &[With
     }
 
     checker.diagnostics.push(Diagnostic::new(
-        violations::NoAssertRaisesException,
+        NoAssertRaisesException,
         Range::from_located(stmt),
     ));
 }
