+++ src/rules/flake8_bugbear/rules/cached_instance_method.rs
@@ -1,9 +1,22 @@
-use rustpython_ast::{Expr, ExprKind};
-
 use crate::ast::types::{Range, ScopeKind};
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Expr, ExprKind};
+
+define_violation!(
+    pub struct CachedInstanceMethod;
+);
+impl Violation for CachedInstanceMethod {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!(
+            "Use of `functools.lru_cache` or `functools.cache` on methods can lead to memory leaks"
+        )
+    }
+}
 
 fn is_cache_func(checker: &Checker, expr: &Expr) -> bool {
     checker.resolve_call_path(expr).map_or(false, |call_path| {
@@ -35,7 +48,7 @@ pub fn cached_instance_method(checker: &mut Checker, decorator_list: &[Expr]) {
             },
         ) {
             checker.diagnostics.push(Diagnostic::new(
-                violations::CachedInstanceMethod,
+                CachedInstanceMethod,
                 Range::from_located(decorator),
             ));
         }
