+++ src/rules/flake8_bugbear/rules/cannot_raise_literal.rs
@@ -1,9 +1,20 @@
-use rustpython_ast::{Expr, ExprKind};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Expr, ExprKind};
+
+define_violation!(
+    pub struct CannotRaiseLiteral;
+);
+impl Violation for CannotRaiseLiteral {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Cannot raise a literal. Did you intend to return it or raise an Exception?")
+    }
+}
 
 /// B016
 pub fn cannot_raise_literal(checker: &mut Checker, expr: &Expr) {
@@ -11,7 +22,7 @@ pub fn cannot_raise_literal(checker: &mut Checker, expr: &Expr) {
         return;
     };
     checker.diagnostics.push(Diagnostic::new(
-        violations::CannotRaiseLiteral,
+        CannotRaiseLiteral,
         Range::from_located(expr),
     ));
 }
