+++ src/rules/flake8_bugbear/rules/f_string_docstring.rs
@@ -1,9 +1,23 @@
-use rustpython_ast::{ExprKind, Stmt, StmtKind};
-
 use crate::ast::helpers;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{ExprKind, Stmt, StmtKind};
+
+define_violation!(
+    pub struct FStringDocstring;
+);
+impl Violation for FStringDocstring {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!(
+            "f-string used as docstring. This will be interpreted by python as a joined string \
+             rather than a docstring."
+        )
+    }
+}
 
 /// B021
 pub fn f_string_docstring(checker: &mut Checker, body: &[Stmt]) {
@@ -17,7 +31,7 @@ pub fn f_string_docstring(checker: &mut Checker, body: &[Stmt]) {
         return;
     };
     checker.diagnostics.push(Diagnostic::new(
-        violations::FStringDocstring,
+        FStringDocstring,
         helpers::identifier_range(stmt, checker.locator),
     ));
 }
