+++ src/rules/flake8_bugbear/rules/getattr_with_constant.rs
@@ -1,14 +1,31 @@
-use rustpython_ast::{Constant, Expr, ExprContext, ExprKind, Location};
-
 use crate::ast::helpers::unparse_expr;
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::fix::Fix;
 use crate::python::identifiers::{is_identifier, is_mangled_private};
 use crate::python::keyword::KWLIST;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::AlwaysAutofixableViolation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Constant, Expr, ExprContext, ExprKind, Location};
 
+define_violation!(
+    pub struct GetAttrWithConstant;
+);
+impl AlwaysAutofixableViolation for GetAttrWithConstant {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!(
+            "Do not call `getattr` with a constant attribute value. It is not any safer than \
+             normal property access."
+        )
+    }
+
+    fn autofix_title(&self) -> String {
+        "Replace `getattr` with attribute access".to_string()
+    }
+}
 fn attribute(value: &Expr, attr: &str) -> Expr {
     Expr::new(
         Location::default(),
@@ -45,8 +62,7 @@ pub fn getattr_with_constant(checker: &mut Checker, expr: &Expr, func: &Expr, ar
         return;
     }
 
-    let mut diagnostic =
-        Diagnostic::new(violations::GetAttrWithConstant, Range::from_located(expr));
+    let mut diagnostic = Diagnostic::new(GetAttrWithConstant, Range::from_located(expr));
 
     if checker.patch(diagnostic.kind.rule()) {
         diagnostic.amend(Fix::replacement(
