+++ src/rules/flake8_bugbear/rules/mod.rs
@@ -1,29 +1,45 @@
-pub use abstract_base_class::abstract_base_class;
-pub use assert_false::assert_false;
-pub use assert_raises_exception::assert_raises_exception;
-pub use assignment_to_os_environ::assignment_to_os_environ;
-pub use cached_instance_method::cached_instance_method;
-pub use cannot_raise_literal::cannot_raise_literal;
-pub use duplicate_exceptions::duplicate_exceptions;
-pub use f_string_docstring::f_string_docstring;
-pub use function_call_argument_default::function_call_argument_default;
-pub use function_uses_loop_variable::function_uses_loop_variable;
-pub use getattr_with_constant::getattr_with_constant;
-pub use jump_statement_in_finally::jump_statement_in_finally;
-pub use loop_variable_overrides_iterator::loop_variable_overrides_iterator;
-pub use mutable_argument_default::mutable_argument_default;
-pub use raise_without_from_inside_except::raise_without_from_inside_except;
-pub use redundant_tuple_in_exception_handler::redundant_tuple_in_exception_handler;
-pub use setattr_with_constant::setattr_with_constant;
-pub use star_arg_unpacking_after_keyword_arg::star_arg_unpacking_after_keyword_arg;
-pub use strip_with_multi_characters::strip_with_multi_characters;
-pub use unary_prefix_increment::unary_prefix_increment;
-pub use unreliable_callable_check::unreliable_callable_check;
-pub use unused_loop_control_variable::unused_loop_control_variable;
-pub use useless_comparison::useless_comparison;
-pub use useless_contextlib_suppress::useless_contextlib_suppress;
-pub use useless_expression::useless_expression;
-pub use zip_without_explicit_strict::zip_without_explicit_strict;
+pub use abstract_base_class::{
+    abstract_base_class, AbstractBaseClassWithoutAbstractMethod,
+    EmptyMethodWithoutAbstractDecorator,
+};
+pub use assert_false::{assert_false, DoNotAssertFalse};
+pub use assert_raises_exception::{assert_raises_exception, NoAssertRaisesException};
+pub use assignment_to_os_environ::{assignment_to_os_environ, AssignmentToOsEnviron};
+pub use cached_instance_method::{cached_instance_method, CachedInstanceMethod};
+pub use cannot_raise_literal::{cannot_raise_literal, CannotRaiseLiteral};
+pub use duplicate_exceptions::{
+    duplicate_exceptions, DuplicateHandlerException, DuplicateTryBlockException,
+};
+pub use f_string_docstring::{f_string_docstring, FStringDocstring};
+pub use function_call_argument_default::{
+    function_call_argument_default, FunctionCallArgumentDefault,
+};
+pub use function_uses_loop_variable::{function_uses_loop_variable, FunctionUsesLoopVariable};
+pub use getattr_with_constant::{getattr_with_constant, GetAttrWithConstant};
+pub use jump_statement_in_finally::{jump_statement_in_finally, JumpStatementInFinally};
+pub use loop_variable_overrides_iterator::{
+    loop_variable_overrides_iterator, LoopVariableOverridesIterator,
+};
+pub use mutable_argument_default::{mutable_argument_default, MutableArgumentDefault};
+pub use raise_without_from_inside_except::{
+    raise_without_from_inside_except, RaiseWithoutFromInsideExcept,
+};
+pub use redundant_tuple_in_exception_handler::{
+    redundant_tuple_in_exception_handler, RedundantTupleInExceptionHandler,
+};
+
+pub use setattr_with_constant::{setattr_with_constant, SetAttrWithConstant};
+pub use star_arg_unpacking_after_keyword_arg::{
+    star_arg_unpacking_after_keyword_arg, StarArgUnpackingAfterKeywordArg,
+};
+pub use strip_with_multi_characters::{strip_with_multi_characters, StripWithMultiCharacters};
+pub use unary_prefix_increment::{unary_prefix_increment, UnaryPrefixIncrement};
+pub use unreliable_callable_check::{unreliable_callable_check, UnreliableCallableCheck};
+pub use unused_loop_control_variable::{unused_loop_control_variable, UnusedLoopControlVariable};
+pub use useless_comparison::{useless_comparison, UselessComparison};
+pub use useless_contextlib_suppress::{useless_contextlib_suppress, UselessContextlibSuppress};
+pub use useless_expression::{useless_expression, UselessExpression};
+pub use zip_without_explicit_strict::{zip_without_explicit_strict, ZipWithoutExplicitStrict};
 
 mod abstract_base_class;
 mod assert_false;
