+++ src/rules/flake8_bugbear/rules/strip_with_multi_characters.rs
@@ -1,10 +1,21 @@
-use itertools::Itertools;
-use rustpython_ast::{Constant, Expr, ExprKind};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use itertools::Itertools;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Constant, Expr, ExprKind};
+
+define_violation!(
+    pub struct StripWithMultiCharacters;
+);
+impl Violation for StripWithMultiCharacters {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Using `.strip()` with multi-character strings is misleading the reader")
+    }
+}
 
 /// B005
 pub fn strip_with_multi_characters(checker: &mut Checker, expr: &Expr, func: &Expr, args: &[Expr]) {
@@ -27,7 +38,7 @@ pub fn strip_with_multi_characters(checker: &mut Checker, expr: &Expr, func: &Ex
 
     if value.len() > 1 && value.chars().unique().count() != value.len() {
         checker.diagnostics.push(Diagnostic::new(
-            violations::StripWithMultiCharacters,
+            StripWithMultiCharacters,
             Range::from_located(expr),
         ));
     }
