+++ src/rules/flake8_bugbear/rules/unary_prefix_increment.rs
@@ -17,12 +17,23 @@
 //! n += 1
 //! ```
 
-use rustpython_ast::{Expr, ExprKind, Unaryop};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Expr, ExprKind, Unaryop};
+
+define_violation!(
+    pub struct UnaryPrefixIncrement;
+);
+impl Violation for UnaryPrefixIncrement {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Python does not support the unary prefix increment")
+    }
+}
 
 /// B002
 pub fn unary_prefix_increment(checker: &mut Checker, expr: &Expr, op: &Unaryop, operand: &Expr) {
@@ -36,7 +47,7 @@ pub fn unary_prefix_increment(checker: &mut Checker, expr: &Expr, op: &Unaryop, 
         return;
     }
     checker.diagnostics.push(Diagnostic::new(
-        violations::UnaryPrefixIncrement,
+        UnaryPrefixIncrement,
         Range::from_located(expr),
     ));
 }
