+++ src/rules/flake8_bugbear/rules/unreliable_callable_check.rs
@@ -1,9 +1,23 @@
-use rustpython_ast::{Constant, Expr, ExprKind};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Constant, Expr, ExprKind};
+
+define_violation!(
+    pub struct UnreliableCallableCheck;
+);
+impl Violation for UnreliableCallableCheck {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!(
+            " Using `hasattr(x, '__call__')` to test if x is callable is unreliable. Use \
+             `callable(x)` for consistent results."
+        )
+    }
+}
 
 /// B004
 pub fn unreliable_callable_check(checker: &mut Checker, expr: &Expr, func: &Expr, args: &[Expr]) {
@@ -27,7 +41,7 @@ pub fn unreliable_callable_check(checker: &mut Checker, expr: &Expr, func: &Expr
         return;
     }
     checker.diagnostics.push(Diagnostic::new(
-        violations::UnreliableCallableCheck,
+        UnreliableCallableCheck,
         Range::from_located(expr),
     ));
 }
