+++ src/rules/flake8_bugbear/rules/useless_comparison.rs
@@ -1,14 +1,28 @@
-use rustpython_ast::{Expr, ExprKind};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Expr, ExprKind};
+
+define_violation!(
+    pub struct UselessComparison;
+);
+impl Violation for UselessComparison {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!(
+            "Pointless comparison. This comparison does nothing but waste CPU instructions. \
+             Either prepend `assert` or remove it."
+        )
+    }
+}
 
 pub fn useless_comparison(checker: &mut Checker, expr: &Expr) {
     if matches!(expr.node, ExprKind::Compare { .. }) {
         checker.diagnostics.push(Diagnostic::new(
-            violations::UselessComparison,
+            UselessComparison,
             Range::from_located(expr),
         ));
     }
