+++ src/rules/flake8_bugbear/rules/useless_contextlib_suppress.rs
@@ -1,9 +1,23 @@
-use rustpython_ast::Expr;
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::Expr;
+
+define_violation!(
+    pub struct UselessContextlibSuppress;
+);
+impl Violation for UselessContextlibSuppress {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!(
+            "No arguments passed to `contextlib.suppress`. No exceptions will be suppressed and \
+             therefore this context manager is redundant"
+        )
+    }
+}
 
 /// B005
 pub fn useless_contextlib_suppress(checker: &mut Checker, expr: &Expr, args: &[Expr]) {
@@ -13,7 +27,7 @@ pub fn useless_contextlib_suppress(checker: &mut Checker, expr: &Expr, args: &[E
         })
     {
         checker.diagnostics.push(Diagnostic::new(
-            violations::UselessContextlibSuppress,
+            UselessContextlibSuppress,
             Range::from_located(expr),
         ));
     }
