+++ src/rules/flake8_bugbear/rules/useless_expression.rs
@@ -1,9 +1,20 @@
-use rustpython_ast::{Constant, ExprKind, Stmt, StmtKind};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Constant, ExprKind, Stmt, StmtKind};
+
+define_violation!(
+    pub struct UselessExpression;
+);
+impl Violation for UselessExpression {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Found useless expression. Either assign it to a variable or remove it.")
+    }
+}
 
 // B018
 pub fn useless_expression(checker: &mut Checker, body: &[Stmt]) {
@@ -12,7 +23,7 @@ pub fn useless_expression(checker: &mut Checker, body: &[Stmt]) {
             match &value.node {
                 ExprKind::List { .. } | ExprKind::Dict { .. } | ExprKind::Set { .. } => {
                     checker.diagnostics.push(Diagnostic::new(
-                        violations::UselessExpression,
+                        UselessExpression,
                         Range::from_located(value),
                     ));
                 }
@@ -20,7 +31,7 @@ pub fn useless_expression(checker: &mut Checker, body: &[Stmt]) {
                     Constant::Str { .. } | Constant::Ellipsis => {}
                     _ => {
                         checker.diagnostics.push(Diagnostic::new(
-                            violations::UselessExpression,
+                            UselessExpression,
                             Range::from_located(value),
                         ));
                     }
