+++ src/rules/flake8_bugbear/rules/zip_without_explicit_strict.rs
@@ -1,9 +1,20 @@
-use rustpython_ast::{Expr, ExprKind, Keyword};
-
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+use rustpython_ast::{Expr, ExprKind, Keyword};
+
+define_violation!(
+    pub struct ZipWithoutExplicitStrict;
+);
+impl Violation for ZipWithoutExplicitStrict {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("`zip()` without an explicit `strict=` parameter")
+    }
+}
 
 /// B905
 pub fn zip_without_explicit_strict(
@@ -24,7 +35,7 @@ pub fn zip_without_explicit_strict(
             })
         {
             checker.diagnostics.push(Diagnostic::new(
-                violations::ZipWithoutExplicitStrict,
+                ZipWithoutExplicitStrict,
                 Range::from_located(expr),
             ));
         }
