+++ src/violations.rs
@@ -988,399 +988,6 @@ impl Violation for BuiltinAttributeShadowing {
     }
 }
 
-// flake8-bugbear
-
-define_violation!(
-    pub struct UnaryPrefixIncrement;
-);
-impl Violation for UnaryPrefixIncrement {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Python does not support the unary prefix increment")
-    }
-}
-
-define_violation!(
-    pub struct AssignmentToOsEnviron;
-);
-impl Violation for AssignmentToOsEnviron {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Assigning to `os.environ` doesn't clear the environment")
-    }
-}
-
-define_violation!(
-    pub struct UnreliableCallableCheck;
-);
-impl Violation for UnreliableCallableCheck {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!(
-            " Using `hasattr(x, '__call__')` to test if x is callable is unreliable. Use \
-             `callable(x)` for consistent results."
-        )
-    }
-}
-
-define_violation!(
-    pub struct StripWithMultiCharacters;
-);
-impl Violation for StripWithMultiCharacters {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Using `.strip()` with multi-character strings is misleading the reader")
-    }
-}
-
-define_violation!(
-    pub struct MutableArgumentDefault;
-);
-impl Violation for MutableArgumentDefault {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Do not use mutable data structures for argument defaults")
-    }
-}
-
-define_violation!(
-    pub struct UnusedLoopControlVariable {
-        /// The name of the loop control variable.
-        pub name: String,
-        /// Whether the variable is safe to rename. A variable is unsafe to
-        /// rename if it _might_ be used by magic control flow (e.g.,
-        /// `locals()`).
-        pub safe: bool,
-    }
-);
-impl Violation for UnusedLoopControlVariable {
-    const AUTOFIX: Option<AutofixKind> = Some(AutofixKind::new(Availability::Always));
-
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let UnusedLoopControlVariable { name, safe } = self;
-        if *safe {
-            format!("Loop control variable `{name}` not used within loop body")
-        } else {
-            format!("Loop control variable `{name}` may not be used within loop body")
-        }
-    }
-
-    fn autofix_title_formatter(&self) -> Option<fn(&Self) -> String> {
-        let UnusedLoopControlVariable { safe, .. } = self;
-        if *safe {
-            Some(|UnusedLoopControlVariable { name, .. }| {
-                format!("Rename unused `{name}` to `_{name}`")
-            })
-        } else {
-            None
-        }
-    }
-}
-
-define_violation!(
-    pub struct FunctionCallArgumentDefault {
-        pub name: Option<String>,
-    }
-);
-impl Violation for FunctionCallArgumentDefault {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let FunctionCallArgumentDefault { name } = self;
-        if let Some(name) = name {
-            format!("Do not perform function call `{name}` in argument defaults")
-        } else {
-            format!("Do not perform function call in argument defaults")
-        }
-    }
-}
-
-define_violation!(
-    pub struct GetAttrWithConstant;
-);
-impl AlwaysAutofixableViolation for GetAttrWithConstant {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!(
-            "Do not call `getattr` with a constant attribute value. It is not any safer than \
-             normal property access."
-        )
-    }
-
-    fn autofix_title(&self) -> String {
-        "Replace `getattr` with attribute access".to_string()
-    }
-}
-
-define_violation!(
-    pub struct SetAttrWithConstant;
-);
-impl AlwaysAutofixableViolation for SetAttrWithConstant {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!(
-            "Do not call `setattr` with a constant attribute value. It is not any safer than \
-             normal property access."
-        )
-    }
-
-    fn autofix_title(&self) -> String {
-        "Replace `setattr` with assignment".to_string()
-    }
-}
-
-define_violation!(
-    pub struct DoNotAssertFalse;
-);
-impl AlwaysAutofixableViolation for DoNotAssertFalse {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Do not `assert False` (`python -O` removes these calls), raise `AssertionError()`")
-    }
-
-    fn autofix_title(&self) -> String {
-        "Replace `assert False`".to_string()
-    }
-}
-
-define_violation!(
-    pub struct JumpStatementInFinally {
-        pub name: String,
-    }
-);
-impl Violation for JumpStatementInFinally {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let JumpStatementInFinally { name } = self;
-        format!("`{name}` inside `finally` blocks cause exceptions to be silenced")
-    }
-}
-
-define_violation!(
-    pub struct RedundantTupleInExceptionHandler {
-        pub name: String,
-    }
-);
-impl AlwaysAutofixableViolation for RedundantTupleInExceptionHandler {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let RedundantTupleInExceptionHandler { name } = self;
-        format!(
-            "A length-one tuple literal is redundant. Write `except {name}` instead of `except \
-             ({name},)`."
-        )
-    }
-
-    fn autofix_title(&self) -> String {
-        let RedundantTupleInExceptionHandler { name } = self;
-        format!("Replace with `except {name}`")
-    }
-}
-
-define_violation!(
-    pub struct DuplicateHandlerException {
-        pub names: Vec<String>,
-    }
-);
-impl AlwaysAutofixableViolation for DuplicateHandlerException {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let DuplicateHandlerException { names } = self;
-        if names.len() == 1 {
-            let name = &names[0];
-            format!("Exception handler with duplicate exception: `{name}`")
-        } else {
-            let names = names.iter().map(|name| format!("`{name}`")).join(", ");
-            format!("Exception handler with duplicate exceptions: {names}")
-        }
-    }
-
-    fn autofix_title(&self) -> String {
-        "De-duplicate exceptions".to_string()
-    }
-}
-
-define_violation!(
-    pub struct UselessComparison;
-);
-impl Violation for UselessComparison {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!(
-            "Pointless comparison. This comparison does nothing but waste CPU instructions. \
-             Either prepend `assert` or remove it."
-        )
-    }
-}
-
-define_violation!(
-    pub struct CannotRaiseLiteral;
-);
-impl Violation for CannotRaiseLiteral {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Cannot raise a literal. Did you intend to return it or raise an Exception?")
-    }
-}
-
-define_violation!(
-    pub struct NoAssertRaisesException;
-);
-impl Violation for NoAssertRaisesException {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`assertRaises(Exception)` should be considered evil")
-    }
-}
-
-define_violation!(
-    pub struct UselessExpression;
-);
-impl Violation for UselessExpression {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Found useless expression. Either assign it to a variable or remove it.")
-    }
-}
-
-define_violation!(
-    pub struct CachedInstanceMethod;
-);
-impl Violation for CachedInstanceMethod {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!(
-            "Use of `functools.lru_cache` or `functools.cache` on methods can lead to memory leaks"
-        )
-    }
-}
-
-define_violation!(
-    pub struct LoopVariableOverridesIterator {
-        pub name: String,
-    }
-);
-impl Violation for LoopVariableOverridesIterator {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let LoopVariableOverridesIterator { name } = self;
-        format!("Loop control variable `{name}` overrides iterable it iterates")
-    }
-}
-
-define_violation!(
-    pub struct FStringDocstring;
-);
-impl Violation for FStringDocstring {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!(
-            "f-string used as docstring. This will be interpreted by python as a joined string \
-             rather than a docstring."
-        )
-    }
-}
-
-define_violation!(
-    pub struct UselessContextlibSuppress;
-);
-impl Violation for UselessContextlibSuppress {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!(
-            "No arguments passed to `contextlib.suppress`. No exceptions will be suppressed and \
-             therefore this context manager is redundant"
-        )
-    }
-}
-
-define_violation!(
-    pub struct FunctionUsesLoopVariable {
-        pub name: String,
-    }
-);
-impl Violation for FunctionUsesLoopVariable {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let FunctionUsesLoopVariable { name } = self;
-        format!("Function definition does not bind loop variable `{name}`")
-    }
-}
-
-define_violation!(
-    pub struct AbstractBaseClassWithoutAbstractMethod {
-        pub name: String,
-    }
-);
-impl Violation for AbstractBaseClassWithoutAbstractMethod {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let AbstractBaseClassWithoutAbstractMethod { name } = self;
-        format!("`{name}` is an abstract base class, but it has no abstract methods")
-    }
-}
-
-define_violation!(
-    pub struct DuplicateTryBlockException {
-        pub name: String,
-    }
-);
-impl Violation for DuplicateTryBlockException {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let DuplicateTryBlockException { name } = self;
-        format!("try-except block with duplicate exception `{name}`")
-    }
-}
-
-define_violation!(
-    pub struct StarArgUnpackingAfterKeywordArg;
-);
-impl Violation for StarArgUnpackingAfterKeywordArg {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Star-arg unpacking after a keyword argument is strongly discouraged")
-    }
-}
-
-define_violation!(
-    pub struct EmptyMethodWithoutAbstractDecorator {
-        pub name: String,
-    }
-);
-impl Violation for EmptyMethodWithoutAbstractDecorator {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let EmptyMethodWithoutAbstractDecorator { name } = self;
-        format!(
-            "`{name}` is an empty method in an abstract base class, but has no abstract decorator"
-        )
-    }
-}
-
-define_violation!(
-    pub struct RaiseWithoutFromInsideExcept;
-);
-impl Violation for RaiseWithoutFromInsideExcept {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!(
-            "Within an except clause, raise exceptions with `raise ... from err` or `raise ... \
-             from None` to distinguish them from errors in exception handling"
-        )
-    }
-}
-
-define_violation!(
-    pub struct ZipWithoutExplicitStrict;
-);
-impl Violation for ZipWithoutExplicitStrict {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`zip()` without an explicit `strict=` parameter")
-    }
-}
-
 // flake8-blind-except
 
 define_violation!(
